requirements.txt
Flask==3.0.3
python-dotenv
pdfplumber
google-generativeai
